Refactor: move channel_reference to new module
