+++ response/slack/reference_utils.py
@@ -1,0 +1,4 @@
+def channel_reference(channel_id):
+    if channel_id is None:
+        return None
+    return f"<#{channel_id}>"
